Exception docstrings and too negative charge states
Checking the docstrings to make sure that the exceptions listed
actually correspond to the exceptions that are called in the code.

Making some of these functions raise a UserWarning when the charge
state is -4 or below.  When I asked my upstairs neighbor (a chemist)
what the most negative charge that she has ever run across, she said
-2 and that even that required special conditions.  We could thus
possibly change this to -3 or below, but there could be some instances
where a -3 charge state could be what a user wants.