Create _check_quantity plus tests
This method checks that an object is a Quantity with acceptable units
and valid numerical values.  Optional keyword arguments provide some
flexibility, including options to allow or disallow Quantities with
negative values (e.g., for densities), complex numbers, or infinities.
This method is intended to simplify code for calculating things like
plasma parameters, since this would otherwise require significant
duplication of code.

An alternative is the quantity_input decorator in astropy.  However,
quantity_input does not check for NaNs or other constraints on values.
Moreover, there are some functions in parameters.py that allow units
of two different types (e.g., finding a gyroradius can be done using
either a perpendicular velocity, or a temperature if you want to use a
thermal velocity).

In the future we might want to turn this into a decorator, or to
propose upstream changes to quantity_input.