Creating _check_relativistic and tests
Creating a method that will check whether or not a Quantity
representing a velocity is relativstic.  This method is primarily for
use in checking plasma parameter calculations to see if we end up with
Alfven speeds exceeding 0.1*c, and raising UserWarnings if it does.
The Lorentz factor for 0.1*c is about 1.005.

A possibility for refactoring in the future would be to create a
decorator that will check whether or not the velocity returned by a
function is relativistic.  This would probably be more Pythonic, but
on the other hand, I'm getting kind of sleepy.

Also: renaming quantities.py to checks.py.