Significant modifications to parameters.py + tests
Updating parameters.py to use _check_quantity and _check_relativistic
from utils subpackage.  The first checks to make sure an object is a
quantity with the correct units and acceptable numerical values.  The
second checks velocities to make sure they're not relativistic.  This
change greatly reduces the repetition of the same checks being done in
different functions, but should probably be changed into decorators at
some point after the universe has expanded a little more.

I removed an error in the calculation of gyroradii and inertial
lengths which were of the form of a velocity divided by an angular
frequency.  I used the dimensionless_angles equivalency from
astropy.units, which treats radians as 1.  Hence, the results were a
factor of 2*pi off.  The new equivalency list is [(u.cy/u.s, u.Hz)]
which yields the intended behavior.  Radians are weird.

I updated a lot of docstrings.  These still need to be checked.
Additionally, it would be great to have an understandable discussion
of the physics in each of the routines.  I still need to add upper &
lower hybrid frequencies.  The tests for parameters.py still need to
be extended and updated.