Modifying arguments to functions in parameters.py
Changing electron_gyroradius so that instead of Vperp_or_Te as an argument, there are separate Vperp and T_e keyword arguments.  Allowing there to be one positional argument with units appropriate for Vperp and T_e which would then be assigned to Vperp or T_e.

As was suggested, if only a number is provided rather than a Quantity, a UserWarning will be raised and units will be assigned as expected (presumably SI units).

Updating some tests to allow these changes.  More tests need to be written (e.g., that UserWarnings are raised when only numbers are inputted).