Updating test on _check_quantity
Tests on my last commit failed because I forgot to update the tests on _check_quantity, which now accepts numbers and attaches units to them in cases where it is unambiguous.  Instead of an error, this should now raise a UserWarning.

With this commit, the universe should be safe once again.  But for how long?