Allowing arguments to default to SI values in parameters.py
I added tests to check that if a number rather than an astropy.units.Quantity is provided, the SI units will be assumed and a UserWarning will be raised.  I updated ion_gyroradius to allow Vperp and T_i as inputs directly and removed Vperp_or_Ti as a keyword.  I updated the docstrings for ion_gyroradius and electron_gyroradius.