Allow ion_mass('e') to return m_e
In deBroglie_wavelength() it will often be necessary to get the
electron mass.  This is a bit of a hack, but allowing ion_mass to
return the electron mass makes the code a bit easier.  We'll need to
rework the atomic routines a bit since using ion_mass() to get the
electron mass is a bit awkward.