Near completion of code for Coulomb_logarithm
I think I got Coulomb_logarithm to work, but at the cost of no longer
believing anything related to plasma transport theory ever again.  My
goal was to make the code as physics-based and as close to a
derivation as possible, so that users will understand why they should
never believe in plasma transport ever again either.

The choice of inner de Broglie wavelength using the reduced mass seems
reasonable, because for an electron-ion pair then the reduced mass
reduces to approximately the electron mass which corresponds to the
electron de Broglie wavelength which will be longer than the ion de
Broglie wavelength.

I put a bunch of (sometimes contradictory) references in a comment on
pull request #60 in the PlasmaPy GitHub repository.

I still have to finish documentation and add tests.  It would be good
to compare to the values in Table 5.1 of Spitzer (1962).