Add patent clause to license
The Open Source Initiative (OSI) recently approved a new license that
consists of the BSD 2-clause license with an added patent clause.  The
purpose of the patent clause is to give users of PlasmaPy the right to
continue to use the software even if a contributor patents their
contribution.  This commit adds the patent clause to our BSD 3-clause
license to ensure that future users have the right to continue to use
the package even if part of it gets patented.

This license takes effect for all future contributions to and versions
of PlasmaPy.  Earlier versions of PlasmaPy can continue to be
distributed under the previous license.  I consent to granting this
patent clause for my prior contributions.

BSD-2-Clause-Patent license: https://opensource.org/licenses/BSDplusPatent