Refactor nuclear_reaction_energy
Originally, the function nuclear_reaction_energy required a string
containing the nuclear reaction as the sole argument.  In issue #110,
@jessicalostinspace suggested that the inputs be lists of the
reactants and products in the reaction which will make it easier to
iterate over different reactions.  This pull request implements this
suggestion.  I did require that reactants and products be keyword-only
arguments because there is potential for the two lists to get mixed
up, and the sign of the reaction is important.

I decided to keep the old API which has the argument being a string
representing the function, but with some ambivalence.  The advantage
of this is that it is much more human readable if we are doing a small
number of reactions.  The disadvantage is that it makes the
implementation a bit more complex.  I broke the code into different
functions so that the flow of the main part of the function is
smoother.

Electrons and positrons are now allowed in the reactions since they
are important for beta decays.