Use ion_mass in nuclear_binding_energy
Prior to this, nuclear_binding_energy used isotope_mass as the mass of
the nucleus.  However, isotope_mass includes the masses of the bound
electrons, which should not be included in the nuclear binding energy.
This function now uses ion_mass and assumes that the isotope is fully
ionized.

There is at least one error at perhaps the 1e-5 level.  For Fe-56, the
function gave 492259.578 keV whereas the corresponding result on
Wikipedia is 492253.892 keV for a difference of 5.686 keV.  Plasma
physicists rarely need this level of accuracy in practice, but we
might still be able to get more accurate values for nuclide masses and
binding energies for any applications that need it.