Add tests for nuclear_reaction_energy
The tests are mostly for when nuclear_reaction_energy accepts the
reactants and products keyword only arguments.  I added tests for the
CNO cycle for the precision given in Wikipedia and some additional
examples from the World Wide Web (TM).  The tests for exceptions with
keyword arguments now include violations of charge conservation and
nucleon number conservation.