Remove clause from atomic tests to ignore warnings
When I changed some functionality earlier to have warnings when
assumptions were made, I added these lines in so that the warnings
were ignored.  I reverted those commits, so these lines are no longer
necessary.

If we do this again, @StanczakDominik made a suggestion:: "Perhaps
this could be simply replaced with two layers of pytest.raises and
pytest.warns?"