Refactor dependency checks in __init__.py
Prior to this, there were separate checks for NumPy and Astropy to
make sure that they were at least the minimum version.  This commit
consolidates those two checks into _check_versions where the argument
of that function is a dictionary with package names as keys and
minimum version strings as the corresponding items.  There's also a
separate _check_python with a separate __minimum_python_version__
variable which makes things easier than including that
__check_versions.

By reducing the number of lines in this file, I'm pretty sure this
will make coveralls grumpy because the coverage percentage will go
down...