Update Python supported version check in __init__.py
A problem with a few of the previous commits is that the appropriate
InputError was not being raised when PlasmaPy was being imported from
Python 2.  This commit removes _check_python from the utils
subdirectory and puts the functionality back into the main
__init__.py.  This is similar to what Astropy does.

I removed test_check_python from test_input_helpers.py and shortened a
function name for PEP 8 compliance.  For the moment, I did not run
`del _split_versions` at the end of __init__.py because it is needed
in test_import_helpers.py.