Add atomic subpackage exceptions and warnings
I added AtomicError, IonError, IsotopeError, ElementError, and
ParticleError.  AtomicError is subclassed from PlasmaPy Error, and the
following exceptions in the list are presently subclassed from the
previous exception, though this will probably need to be thought about
more carefully.

I also added "pass" to each of the exceptions and warnings.  This does
not appear to be needed, but this line will show up in coverage tests
for exceptions and warnings that are not being tested.