Add code to construct _Particles dictionary
Instead of manually inputting data (which can lead to errors and
duplication of code), this commit adds a function that initializes the
dictionary used to store information about particles and antiparticles.

I have some loops like `for thing in everything`.  I originally used
`for particle in particles` but this doesn't account for some of the
things being antiparticles.  It would be helpful to have a more
specific name for the set containing both particles and antiparticles.