Restructure atomic exceptions
ParticleError is now intended to account for all cases where a
particle is inputted but is invalid.

ElementError, IonError, and IsotopeError are now intended to account
for cases where a particle is valid, but is not of the expected type.
These are all now subclasses of a new exception called
UnexpectedParticleError.