Update Particle class and tests
Changed the default symbol of a proton to be 'p+' for consistency with
'e-' for electrons.  Protons are turning out to be a special special
case since they also have element, isotope, and ion information and
therefore cannot be treated like the other special particles.

Drafted a test for the Particle class that would check that the
attributes give expected results and expected exceptions are raised;
this is not yet finished.