Update __eq__ magic method in Particle class
The easiest way to implement __eq__ for the Particle class would
be to simply check if the particle symbol is the same between the
the two instances of the class.  However, a more detailed check
is desirable because it allows for more reliable testing of this
class.  In particular, this dunder method is really helpful in
making sure that particle aliases produce equivalent results
[e.g., that Particle('p') == Particle('proton')].  This method
also checks that if an exception is raised for one instance of
the class, then it also checks that an exception is raised for
the other instance of the class (without actually checking that
the same exact exception is raised).