Update Particle class
I simplified __eq__ and __ne__ by comparing __dict__ between both
objects instead.  To do this, I had to remove the _original_*
parameters which are probably unnecessary anyway.  I cleaned up
reduced mass and updated some of the documentation throughout.  I
added the spin of an alpha particle as a special case, though I am
wondering if it would be better to have a Nuclide subclass also.