Update Particle class
Further implement the is_category method which replaces is_stable and
is_antimatter.  Refactor the code in a few places to slightly clean
things up.  Updated docstrings for some attributes.

The element, isotope, and ion attributes now return None when the
particle does not fall into that category, instead of raising an
exception.  This allows for a very convenient way of testing whether
or not a particle falls into a particular combination of these
categories (e.g., `if particle.element and not particle.isotope:`.