Update Particle class
I implemented ParticleZoo in the place where categories (such as
'lepton', 'fermion', etc.) are assigned.  Previously, I used exec to
execute the strings of the commands that I created.  However, this
does not play well with PyCharm and other static type checkers that
have a harder time interpreting what is happenening in lines that use
exec.

Previously, the private variables in the Particle class were
initialized by using eval on strings such as "self._mass = None".
This also does not play well with PyCharm, etc.  I changed the way
private variables are initialized so that there are instead a bunch of
statements instead of a loop with eval.

I changed a few flags related to coverage reports to better reflect
what I think does not need to be tested as opposed to what we should
know is not being tested.