Update tests and other things in atomic subpackage
I included one temporary fix at the beginning of isotope_symbol.
There is much more work to do to have the Particle class be used
as the core part of atomic functionality, including the creation
of an appropriate decorator that would take atomic arguments and
turn them into an appropriate Particle object.