Fix test failures in atomic.py
I now know why PyCharm doesn't like it when I do the blanket
`except Exception` for exception handling.  The problem in ion_mass
is that an InvalidParticleError was happening, but different things
needed to be done for InvalidParticleError and InvalidElementError.

Most of the code in atomic.py is going to need to be replaced soon
in favor of using the Particle class, so I have a few other
temporary fixes in there that are...inelegant.