Put Particle class attributes in defaultdict
At the suggestion of Stuart Mumford, the attributes stored within each
instance of the Particle class are now within a defaultdict with None
as the default value.  The nice thing about this strategy is that it
eliminates the long list of initializations of separate individual
attributes to None.  This change makes some of the code more
cumbersome, so it might be helpful to clean things up some more.