Update Particle class and @particle_input + tests
These are significant updates to pass tests.

In @particle_input, I added narrative comments.  I also fixed a bug
where an argument that was already a Particle was not checked to make
sure that it fit the required criteria.  I tried to make the error
messages as clear as I could.

In Particle, I fixed a bunch of bugs in order for tests to pass,
especially related to determination of masses, categories, and
exceptions.  I updated some docstrings too.

I changed Particle.__eq__ to compare only the particle symbol to
address some weird problems where two electron Particle objects were
not being determined to be equal.  I suspect that this may have been
because I used a lambda function in the defaultdict declaration which
didn't register as the same in both instances, but I did not figure
this out for certain.  We may wish to do this more carefully in the
future to make sure more of the attributes are equal.