Update atomic.py with Particle and @particle_input
I made a bunch of bug fixes, as well as some (hopefully minor)
backwards incompatible changes.

I did not use the @particle_input category checking in ion_mass
because I wanted this function to be able to be used for positrons
(e.g., in a pair plasma).

I put back a bunch of the _is_electron, etc. helper functions to make
the number of test failures more manageable.  I need to go through and
remove them from here and elsewhere in the code (including in the
physics subpackage for transport and plasma parameter calculations).