Put ion_mass's singly charged assumption back in
Up until now, ion_mass has worked for a bunch of particles that are
not ions.  In particular, if an element or isotope was given with no
charge information, then ion_mass assumed that the particle was singly
ionized.  I regret making the decision to default to this assumption
since "explicit is better than implicit".  It would be better to make
people write the charge information explicitly in an ion's symbol.

The calculation of ion mass by the Particle class does not make the
assumption that a particle is singly ionized, so ion_mass lost this
capability when I switched over to using @particle_input.  I just put
this functionality back in, along with a few solar masses worth of
DeprecationWarnings.

Some of the other test failures were caused by small discrepancies
from when a Constant is used versus a Quantity with the same value.

I also wrote a particle_mass function that is much more general, and
can be used for functions such as thermal_speed that do not require
charged particles.  This still needs docstrings and tests.