Fix deuteron and triton mass discrepancies
Some recent doctest failures happened because the nuclide mass of
deuterium and tritium did not equal the mass of a deuteron or triton,
repectively.  In order to turn these bugs into test cases, I created a
function to test that the nuclide mass equals the mass of the fully
ionized ion of that isotope.  This also checks that the mass and
nuclide mass attributes are equal to each other for both protons and
neutrons.