Add `none_shall_pass` keyword to @particle_input
This keyword, which should be read in the voice of Gandalf as
portrayed by Ian McKellen at the Bridge of Khazad-dûm, will allow None
to pass through in the annotated argument to the decorated function.
This may be useful in functions such as known_isotopes, where an
argument of None would indicate that this function should return all
known isotopes of every element.  Technically this is a misquote,
although practicality beats purity.