Fix bugs in atomic subpackage
I fixed a few lingering problems that I didn't completely fix when I
pulled in the new code from #260 that updated half-lives from NUBASE
2016.  I moved code added to half_life in #260 to Particle.half_life
for consistency with the rest of the changes to atomic.py.

I found an isotope with no half-life data to keep as an example to
make sure a MissingAtomicDataError is raised in those situations.

In some previous commits, I changed the name of the keywords in
Particle.is_category, which also showed up in @particle_input.  I
updated these keywords for atomic tests and in @particle_input in a
few other places that I had missed.  I also fixed some bugs in
is_category that arose when some of the keywords were not set.

In #260