Update type return annotations for Quantities
Previously I had been mixing up return annotations in the atomic
subpackage for functions and methods that returned Quantities.  The
return annotations for some were `-> u.Quantity` whereas others were
like `-> u.kg`.  For functions decorated with @u.quantity_input, it is
appropriate to use the latter return annotation.  I tried decorating
these functions with @u.quantity_input, but this led to errors since
the arguments and keywords are not Quantities themselves.  Instead, I
changed everything in the atomic subpackage to the former possibility
so that the return annotations are `-> u.Quantity` and don't mention
units..