Test Particle.half_life warnings
For isotopes with poorly known half-lives, the corresponding data is
stored as a str in the _Isotopes dictionary.  This commit adds a test
that calling Particle.half_life for such an isotopes leads to a
MissingAtomicDataWarning being issued and a str being returned.

I also removed some probably unnecessary exception handling from
nuclear_reaction_energy.