Add return annotations in test_particle_input
When a function decorated by particle_input has the keywords Z or
mass_numb, the decorator raises an exception when more than one
argument is annotated with Particle in order to avoid ambiguity about
which particle the Z or mass_numb refers to.  However, the failure of
these tests shows that the exception is also unintentionally raised
when the decorated function has a return annotation of Particle.