Begin creating general test function
The goal of this function is to take a tuple with a function; args and
kwargs for that function; and the expected result, warning, or exception
and then test that it has the appropriate behavior.  This function is
likely to be helpful for functions where the args, kwargs, and result
can be represented straightforwardly.  This function should generally
be used in conjunction with @pytest.mark.parametrize.