Adopt morally ambiguous defaults for pydocstyle
I added a list of the common default settings to setup.cfg also for
easy reference.

The general consensus in the discussion on Matrix is that it is more
readable to start the "one-liner" of a docstring on the line after the
""" rather than on the same line as the """, so I set that as the
default.

I also set it so that the "one-liner" can exceed one line, since I've
found there are occasions where even a concise description of the
physics involved in a function or class needs to take up more space
than that.