Update pycodestyle defaults
Whitespace is required around arithmetic operators to improve
readability by making it easier to parse equations.

For the time being I set E241 and E242 so that there has to be no more
than one space after a comma.  There may be times where having things
aligned will make it more readable, so we might wish to revisit this
in the future.

W504 is set so that during a line break, a line should start with the
binary operator (as opposed to the previous line ending with it).