Reformat _Isotopes
Previously, the _Isotopes dictionary in isotopes.py was built using a
helper function that took in the symbol, name, atomic number, mass
number, stability, half-life, and isotopic abundance and then returned
a small dictionary for each isotope that was then put into the big
_Isotopes dictionary.  The motivation behind this strategy was to
reduce the size of the file.  However, a more significant disadvantage
was that it was not very readable since you had to go back and look at
the top of the file to see what each value represented.

I reformatted _Isotopes so that each attribute is on a different line.
I also renamed/shortened some of the keys since that could be done so
without reducing readability (e.g., 'atomic_mass' became 'mass' and
'isotopic_abundance' became 'abundance').  I had to update the keys
elsewhere in the atomic subpackage.