Remove reduced_mass as a Particle class method
The functionality to find reduced mass just before this commit was in
two places: as a method in the Particle class, and as a function in
atomic.py.  Having it as a Particle class method is not the most
natural place for it since it depends on two particles, so I am
deleting Particle.reduced_mass in favor of the standalone reduced_mass
function.