Use dict comprehensions for helper atomic dicts
Previously the dicts to get the atomic symbol from either the element
name or the atomic number were spelled out in full.  However, the same
information was contained in the _Elements dict, so the data was being
repeated.  I changed these to dictionary comprehensions.

This formulation has the advantage that if the atomic number of
hafnium ever changes, we will only need to update it in a single
place.