Update Particle docstrings, tests, and mass units
I added docstring example sections for most of the attributes of the
Particle class, and updated the docstrings in other ways too.

I changed the units for the isotopic mass and standard atomic weight
to kg instead of u for consistency with the other mass calculations.
This way, we are closer to using base SI units.