Add exception handling to test helper function
I added the first draft of functionality for the test helper function
in order for it to handle cases where the expected outcome is an
exception being raised.  I wanted to make sure there are useful error
messages for when no exception is raised, and for when an
unanticipated exception is raised.  This functionality is untested.

I did not use `with pytest.raises()` because it did not have quite the
functionality that I wanted in order to have a useful error message
when and unanticipated exception is raised.  That said, it may end up
being worthwhile to eventually refactor this to use `pytest.raises`.