Add option to cause test helper tests to fail
When tests are passing, the error messages do not show up, which makes
it hard to check that the formatting is working ok.  This option
causes a bunch of the tests to fail, which makes it possible to read
through the error messages to make sure they're showing up ok.