Refactor test_atomic.py with run_test
This is a major refactoring of tests of atomic functions using the
test helper function, run_test.

Previously there were several separate test functions and tables for
each function.  Moreover, for each atomic function there were several
separate test functions and tables to test the results, warnings, and
exceptions.

In this change, there is one table for each atomic function that
includes the positional arguments, optionally the keyword arguments,
and the expected outcome (which may be a warning or exception as well
as the expected result).  These are all run by the same new test
function which calls run_test.

I also added some code to prepend the function in each set of inputs
in the each test table, in order to avoid having to repeat the same
function over and over in each table.

There is also some code to add in inputs for a bunch of different
functions for different exceptions but for the same set of inputs.

It would be helpful to also have a function that tests to make sure
that two calls with equivalent inputs will yield equivalent outputs.
This would be helpful for inputs like 'p+' and 'H-1+' in atomic.

There are still some tests that I need to refactor into the run_test
formulation, and I may see if there are some simpler ways to go about
this.  I am still wondering if an object-oriented approach would end
up being cleaner in the long run.