Remove ion_mass and isotope_mass
The functionality of ion_mass and isotope_mass are subsets of the
functionality of particle_mass, and it doesn't make much sense to have
three separate functions that are designed to do the same thing.  This
commit removes these two functions so that we are left with only the
most general one: particle_mass.

"There should be one-- and preferably only one --obvious way to do
it."

One thing to be aware of is that ion_mass and isotope_mass checked to
make sure that particles were ions or isotopes, whereas particle_mass
does not.  It is better to use @particle_input to test whether or not
a particle meets the requisite criteria for a function.