Halve runtime of classical_transport tests with one simple trick!
So it turns out if you put in a setup_method() into a Test class that's
run by Pytest, that setup_method is executed for every single method call.
This includes all the calculations involved in classical_transport
initialization, and anything else we are doing in preparation for
these tests.

If you just setup_method into a setup_class @classmethod, it's
going to get called once per class - and that means all the variables
in it are only going to be calculated once, and they'll be shared
across every instance of the class.

It was thus the only sane choice to go all out with this.

Reference material: https://docs.pytest.org/en/latest/xunit_setup.html