Use ~ to return antiparticle of special particles
I used the __invert__ dunder method in the Particle class to allow the
unary operator to return the antiparticle of the particle associated
with a Particle instance.  This only works for special particles and
antiparticles, and thus will not work for any atoms or ions except for
protons.  Magic methods are awesome!