Test tuple annotations in @particle_input
This test checks that functions decorated by @particle_input with an
argument that is annotated with (Particle, Particle) will pass through
the expected Particle instances.  In essence, this test checks that
using this tuple annotation will allow @particle_input to behave in
pretty much the same was as having two separate arguments that are
each decorated with Particle.

When setting this test up, we have to be careful because running
Particle('p+') == 'p+' will return True because the __eq__ magic
method for the Particle class allows comparisons with string
representations of particles.