Remove syntax that is not needed in Python 3.6
In Python 3, it is no longer necessary to have "()" for a definition
of a new class; you can just do:

    class SomeClass:
        ...

It is also no longer necessary to start files with a comment that says
"encoding=utf-8".

I updated the code in a couple of places where parentheses were used
or that had "encoding=utf-8".