Test consistency between different requirements files
We have package requirement information in a few different files in
the requirements directory.  There is some duplication of information,
but this duplication is necessary since the package requirements for
documentation tests, code tests, and the package itself differ.  When
we require a new package or update a minimum version, problems can
arise should we change one file but forget to change the others.  I
added some tests to check for common mistakes and inconsistencies for
when we make these updates.  I still need to add a few more to check
different files.