Update exceptions in .coveragerc
Coverage tests should now exclude lines containing
NotImplementedError, ImportError, ModuleNotFoundError, and
SyntaxError.  It should be unlikely that we need to test any of these
exceptions, whether or not we raise them or are trying to catch them.

I also changed tabs to spaces for consistency with most of the rest of
PlasmaPy.  Hopefully this doesn't break the universe or anything. :-)