Add PlasmaPy contributors to credits.rst
This is a first draft of the contributors to PlasmaPy for our version
0.1.0 release.  This involved some detective work!  I found the names
primarily by searching through our git log and then looking at
people's GitHub profiles.  I also added some people who contributed to
the vision statement while it was on Google Docs, and thus do not have
any credits for their commits.

There are some people who made commits but who do not show up on
GitHub.  I did a quick web search and a possible cause may be that
those contributors' PRs were rebased, which caused the person doing
the rebasing to get recorded as making those commits on GitHub.  I do
not know if this is the case here.  When updating this page, we should
make sure to check both the GitHub contributors page as well as the
git log.

Also: I have not yet verified these names, so there may be mistakes!