Significantly update IonizationState
I added a bunch of attributes and methods plus a bunch of checks, and
I think allowed this to be iterated.  I'm still not sure about what
the best names are for the different attributes, and a lot of what is
included in this so far is up in the air.  I'm not sure if we want to
include a way to calculate collisional ionization equilibrium since
that would require a lot of data to be stored by the package somehow,
but that would be really useful and something that would probably be
used a lot by plasma physicists.