Expand IonizationStates class
I added the capability to iterate over charge states, in which case I
used a `collections.namedtuple` with different attributes for the
integer charge, ionic fraction, and ionic symbol since these are the
most relevant values.  It might end up being reasonable to include the
number density in here too, if available.

Prior to this, I had set separate `rtol` and `atol` keyword arguments
and attributes for use with numpy.isclose.  I changed this to a sole
`tol` attribute because (1) for checking the normalization, absolute
and relative tolerances reduce to the same thing because the relative
tolerance is relative to one, and (2) for tiny abundances, the
absolute tolerance will be more relevant than the relative tolerance.
In calls to numpy.isclose, I explicitly set `rtol` equal to zero so
that it's clear that we're dealing with absolute tolerances and not
relative tolerances.

I defaulted `tol` to 1e-15 since at 1e-16 I got errors with
np.isclose.