Have IonizationState.__getitem__ return namedtuple
Before this, I had IonizationState[value] return just a float for the
ionization fraction.  Now it returns a namedtuple with the integer
charge, ionic fraction, and ionic symbol.  This change is because the
integer charge, ionic fraction, and ionic symbol are all likely to be
needed, and to be consistent with the behavior as an iterator.

It may be worth changing this namedtuple (presently named State) into
a class.  There's also a need to create an IonizationStates class that
includes more elements.