Use isinstance on (int, np.integer) instead of int
I noticed that using isinstance(n, int) was returning False when n was
a numpy.int64 object, for example.  On the World Wide Web™ I found we
could instead do isinstance(n, (int, numpy.integer)).  I made this
change in a bunch of atomic modules for Z and mass_numb.

I did not yet change the docstrings or type hints to reflect that
numpy.integer types are allowed, and am debating whether or not it
would be best to do so.  The docstrings and type hints will be more
explicit with allowed types if we say that numpy.integer types are
allowed, but the docstrings and especially the type hints will get
more difficult to read.  It would be helpful if we could make this
notation more compact.  If we end up checking type hints with mypy, we
might need to at least update the type hints.