Rename _citation_bibtex attribute to __citation__
I had created this variable as a private attribute to avoid putting
something package-specific in dunder space.  However, as Cadair
pointed out, it's becoming common usage to use __citation__.  Since
practicality beats purity, we'll go with that.

Obscure Babylon 5 joke from the conversation on Matrix:

  Q: Why did the Vorlon cross the road?
  A: I have always been here.

This joke is released under a CC BY 4.0 license, with the source code
for the joke being released under LGPLv3.