Reorganize citation files once again
I renamed acknowledging.rst to citation.rst in docs/about so that the
name is more clear and in line with what this sort of file is usually
named.

I copied this file to CITATION.rst (while deleting the link) in the
top-level directory so that we have it there too.  The README file now
links to CITATION.rst.