Clean up & add docstrings to utils/roman/roman.py
I added full docstrings with examples to toRoman and fromRoman.  I
removed one of the custom exceptions because it was essentially a
TypeError.

I removed the reference to Dive Into Python from the header because
the website no longer exists.  I updated the link to the Python 2.1.1
license for the roman subpackage.  The license information should stay
in this file.

I had the option of either putting this functionality into an extern
directory or in utils, and because of what it does, it makes more
sense to put it into utils.  I put it as a subsubpackage to try to
help keep things organized.