Add new tests of toRoman and fromRoman
I included all of the Roman numerals from 1 to ~120 because those are
the ones we might possibly need for spectroscopic notation, and it is
quick to test all of them.  All of the code in test_roman.py is
original and not from the roman package on PyPI, which had a few tests
using unittest that would have needed to be converted to using pytest.