Add is_roman_numeral function
The atomic subpackage occasionally needs to check whether or not a
string is a valid Roman numeral.  Previously, atomic used a match
attribute from a regex pattern, but it would be more readable there to
have a function that tests whether or not a string is a valid Roman
numeral.