Make Particle(arg) == Particle(Particle(arg))
This allows all of the arguments in a list to be converted to a
Particle more simply: `[Particle(particle) for particle in particles]`
instead of saying `if not isinstance(particle, Particle) else
particle`.

We had some discussion (see #486) about whether or not the same
instance should be returned, or if a copy should be returned.  The
decision was to create a copy so that it would behave like
numpy.array.  My hesitation is that list comprehensions like the one
above would create lots of Particle instances for the same particle,
when in actuality only one instance of Particle('electron') would be
necessary.  However, there is a chance that returning the same object
in memory would trip users up, so a more rigorous way of doing this is
to just create a copy.

Co-authored-by: Colby Haggerty <colbych@udel.edu>