Add ionize and recombine methods to Particle class
These methods take a Particle object and run __init__ again for the
same element/isotope but for a different charge.  The ionize method
will raise an InvalidIonError if a bare nucleus is attempted to be
ionized.  The recombine method will issue a warning if the charge
becomes negative; however, no exception will be raised in that case.

One behavior that I'm not certain about is both running __init__ again
and returning the new Particle instance.  We may wish to have just one
behavior or the other.  If we have just the in-place
re-initialization, then we would not be able to do comparisons like
Particle("Fe 5+").recombine() == Particle("Fe 4+") which require the
new Particle instance to be returned.